# admin-handoff/docs/openapi.yaml
openapi: 3.1.0
info:
  title: Adventrya Admin API
  version: 0.1.0
  description: Initial production integration contract for the approved Admin prototype.
servers:
  - url: https://api.adventrya.com/v1
    description: Production placeholder
  - url: http://localhost:4000/v1
    description: Local development
security:
  - cookieSession: []
paths:
  /admin/overview:
    get:
      operationId: getAdminOverview
      parameters:
        - $ref: "#/components/parameters/From"
        - $ref: "#/components/parameters/To"
      responses:
        "200":
          description: Operational summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OverviewResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /admin/orders:
    get:
      operationId: listAdminOrders
      parameters:
        - $ref: "#/components/parameters/From"
        - $ref: "#/components/parameters/To"
        - name: query
          in: query
          schema:
            type: string
        - name: purchaseType
          in: query
          schema:
            $ref: "#/components/schemas/PurchaseType"
        - name: paymentStatus
          in: query
          schema:
            type: string
        - name: generationStatus
          in: query
          schema:
            type: string
        - name: printStatus
          in: query
          schema:
            type: string
        - name: deliveryStatus
          in: query
          schema:
            type: string
        - $ref: "#/components/parameters/Cursor"
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: Paginated orders
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/OrderSummary"
                  nextCursor:
                    type: [string, "null"]
...
  /admin/customers:
    get:
      operationId: listAdminCustomers
      parameters:
        - name: query
          in: query
          schema:
            type: string
        - $ref: "#/components/parameters/From"
        - $ref: "#/components/parameters/To"
        - $ref: "#/components/parameters/Cursor"
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: Authorized customer summary list
...
components:
  securitySchemes:
    cookieSession:
      type: apiKey
      in: cookie
      name: adventrya_admin_session
  parameters:
    From:
      name: from
      in: query
      schema:
        type: string
        format: date
    To:
      name: to
      in: query
      schema:
        type: string
        format: date
    Cursor:
      name: cursor
      in: query
      schema:
        type: string
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 25
...
  schemas:
    PurchaseType:
      type: string
      enum: [DIGITAL, PRINTED_DIGITAL, PRINT_UPGRADE]
    Money:
      type: object
      required: [amountMinor, currency]
      properties:
        amountMinor:
          type: integer
        currency:
          type: string
          enum: [GEL]
...
    OverviewResponse:
      type: object
      required: [metrics, attentionItems, recentOrders]
      properties:
        metrics:
          type: object
          additionalProperties: true
        attentionItems:
          type: array
          items:
            type: object
            additionalProperties: true
        recentOrders:
          type: array
          items:
            $ref: "#/components/schemas/OrderSummary"
    OrderSummary:
      type: object
      required:
        - id
        - publicOrderNumber
        - bookTitle
        - parentName
        - purchaseType
        - total
        - paymentStatus
        - generationStatus
        - printStatus
        - deliveryStatus
        - createdAt
      properties:
        id:
          type: string
        publicOrderNumber:
          type: string
        bookTitle:
          type: string
        childDisplayName:
          type: string
        parentName:
          type: string
        parentEmail:
          type: string
          format: email
        parentPhone:
          type: string
        purchaseType:
          $ref: "#/components/schemas/PurchaseType"
        total:
          $ref: "#/components/schemas/Money"
        paymentStatus:
          type: string
        generationStatus:
          type: string
        printStatus:
          type: string
        deliveryStatus:
          type: string
        createdAt:
          type: string
          format: date-time
...
  responses:
    Unauthorized:
      description: Authentication required
    Forbidden:
      description: Permission denied
...